requirements.txt
numpy~=1.24.3
pandas~=1.5.3
tqdm~=4.65.0
torch~=2.0.1
python-docx~=1.0.1
stanza~=1.5.0
nltk~=3.8.1
wheel~=0.41.2
setuptools~=68.2.0
openpyxl~=3.1.5